small refactoring, connected transport seek action

+++ backend/nextcp2-runtime/src/main/java/nextcp/domainmodel/device/services/ITransport.java
@@ -16,7 +16,7 @@ public interface ITransport
 
     void next();
 
-    void seek(int secondsAbsolute);
+    void seek(long secondsAbsolute);
     
     TransportServiceStateDto getCurrentTransportServiceState();
 }
